เหลือไปทำไรเดอร์ต่อ

+++ lib/pages/customer/Order.dart
@@ -18,7 +18,7 @@ import 'package:geocoding/geocoding.dart';
 
 class OrderPage extends StatefulWidget {
   final List<dynamic> mergedMenus;
-  final double deliveryFee;
+  final int deliveryFee;
   final int order_id;
   final int order_status;
   final String? previousPage;
